+++ data/Funky Chicken.qst
@@ -1,0 +1,1 @@
+{'uid': 140240760433489, 'style': 'FILL IN THE BLANK', 'inquiry': 'you are', 'choices': [], 'answer': 'nice'}
