+++ data/quizzes.qz
@@ -1,0 +1,2 @@
+{'uid': 140122028968305, 'name': 'Monkey Business', 'style': 'generic', 'question_file': 'Monkey Business.qst'}
+{'uid': 140240787601008, 'name': 'Funky Chicken', 'style': 'generic', 'question_file': 'Funky Chicken.qst'}
