+++ pages/edit_specific_question.py
@@ -1,0 +1,8 @@
+from pages.page import Page
+
+
+class EditSpecificQuestion(Page):
+
+    def display(self):
+        pass
+
